+++ packages/react-querybuilder/src/hooks/useQueryBuilder.ts
@@ -5,9 +5,15 @@ import type {
   QueryBuilderProps,
   RuleGroupTypeAny,
 } from '../types';
+import type { UseQueryBuilderSchema } from './useQueryBuilderSchema';
 import { useQueryBuilderSchema } from './useQueryBuilderSchema';
 import { useQueryBuilderSetup } from './useQueryBuilderSetup';
 
+/**
+ * Calls {@link useQueryBuilderSetup} to massage the props and prepare basic
+ * update/generate methods, then passes the result to {@link useQueryBuilderSchema}
+ * to prepare and return all values required to render {@link QueryBuilder}.
+ */
 export const useQueryBuilder = <
   RG extends RuleGroupTypeAny,
   F extends FullField,
@@ -15,5 +21,5 @@ export const useQueryBuilder = <
   C extends FullCombinator,
 >(
   props: QueryBuilderProps<RG, F, O, C>
-): ReturnType<typeof useQueryBuilderSchema<RG, F, O, C>> =>
+): UseQueryBuilderSchema<RG, F, O, C> =>
   useQueryBuilderSchema<RG, F, O, C>(props, useQueryBuilderSetup(props));