+++ packages/react-querybuilder/src/utils/mergeTranslations.ts
@@ -4,8 +4,6 @@ import { defaultTranslations } from '../defaults';
 import type { Translations } from '../types';
 import { objectEntries, objectKeys } from './objectUtils';
 
-type RSRSS = Record<string, Record<string, string>>;
-
 /**
  * Merges any number of partial {@link Translations} into a single definition.
  */
@@ -32,7 +30,8 @@ export const mergeTranslation = (
   el: keyof Translations,
   keyPropContextMap: Record<string, [ReactNode, ReactNode]>,
   finalize?: boolean
-): RSRSS | undefined => {
+): Record<string, Record<string, string>> | undefined => {
+  type RSRSS = Record<string, Record<string, string>>;
   const finalKeys = objectEntries(keyPropContextMap)
     .map(([key, [pT, cT]]) => [
       key,