+++ utils/fixJSDOMEnvironment.ts
@@ -1,6 +1,5 @@
 import JSDOMEnvironment from 'jest-environment-jsdom';
 
-// https://github.com/facebook/jest/blob/v29.4.3/website/versioned_docs/version-29.4/Configuration.md#testenvironment-string
 export default class FixJSDOMEnvironment extends JSDOMEnvironment {
   constructor(...args: ConstructorParameters<typeof JSDOMEnvironment>) {
     super(...args);